requirements.txt
pyannote.audio
soundfile 
torch 
torchvision 
torchaudio
libsndfile1
os